Add files via upload
Paula Turmina
artist
born 01.09.1991
location Parai, Rio Grande do Sul, Brasil
www.paulaturmina.com

This is how I imagine the feeling of living and seeing the world under the impacts of global warming of 2°C above pre-industrial levels.
Introducing distortions to the normalized version of planet Earth in 2020 with 2°C above pre-industrial levels, slices of time and space are not fully recognizable but say something about both past and future.
The animal-like hand ripping the canvas' skin speaks of instincts and certain features of different bodies that evolved for protection, such as nails, teeth, thorns, etc. Characteristics that can both protect and harm.
The crawling hand represented in this piece can be seen just about enough to question if that is a human hand or an animal claw. A curved pointed horny nail has no beginning as if it has merged with the skin. Such a feature, that is not recognized as possible in our bodies today, suggests an evolution of the body; whilst the earthy color reveals that this transformation has happened under warmer conditions.
Through painting, I investigate how art-making operates concerning the decolonization of thought and the forming of a different relationship between humans and the earth, animals, and plants that inhabit it. 
The visuals I create are drawn from my personal experience, ecological issues, and Latin American colonial history. I believe craft making is a form of creating change and processing the absurdity of life.

Medium: oil on canvas
Dimensions: 150x180cm
London, 2019